Fix: c_parser ignored scopes with nameless inline struct
The `parse_type` would continue looking for tokens even when the c_type
ended with a `{`. This would mean that the following resulted in an
exception:

```
__struct__ = """
  struct {
    int i;
  } s;
"""
```

`parse_type` now ignores any following tokens if `c_type` ends with an
`{` - meaning the tokens afterwards are not element of the current type
scope.

Signed-off-by: Sophie Tyalie <dev@flowerpot.me>